Separated page files from the components folder into their own folder. Fixed the imports from these changes and updated the routes to use the correct relative path.
